Add EKS, IAM roles, and KeyPair module
This commit introduces the creation and configuration of AWS EKS clusters along with associated IAM roles and security configurations. It also includes the addition of a KeyPair module for Jenkins, updates to the variable and output definitions, and a refactoring of the compute module to use a script for Docker and Jenkins setup.